No padding for build panel
Like any UI change this is debatable. I really don't like the padding on the build panel. I've had this in my `styles.less` for a while, and thought I'd make a PR out of it. I'm sure more feel the same as I do.

Thanks for a great theme. My first choice, btw!